Incorporated lots of changes from master branch.Added -wgl.json files that contain references to lower res textures and assets

+++ desktop/src/gaia/cu9/ari/gaiaorbit/desktop/util/GaiaCatalogFilter.java
@@ -182,8 +182,8 @@ public class GaiaCatalogFilter {
             data_out.writeFloat(s.appmag);
             data_out.writeFloat(s.absmag);
             data_out.writeFloat(s.colorbv);
-            data_out.writeFloat(s.posSph.x);
-            data_out.writeFloat(s.posSph.y);
+            data_out.writeFloat((float) s.posSph.x);
+            data_out.writeFloat((float) s.posSph.y);
             data_out.writeFloat((float) s.pos.len());
             data_out.writeInt(s.id);
         }
